+++ CHANGELOG.md
@@ -4,14 +4,20 @@
 
 ### Fixes
 
+### Chore & Maintenance
+
+### Performance
+
+## 24.3.1
+
+### Fixes
+
 - `[jest-cli]` export functions compatible with `import {default}` ([#8080](https://github.com/facebook/jest/pull/8080))
 - `[jest-worker]`: Fix retries and error notification in workers ([#8079](https://github.com/facebook/jest/pull/8079))
 
 ### Chore & Maintenance
 
 - `[pretty-format]`: Use `react-is` instead of manual `$$typeof` checks ([#8060](https://github.com/facebook/jest/pull/8060))
-
-### Performance
 
 ## 24.3.0